build: decouple tapyrus-genesis/tx/cli/verify from tapyrus_chainstate and leveldb
CheckBlock()/CheckBlockHeader()/isBlockHeightInCoinbase()/FormatStateMessage()
move from validation.cpp into a new checkblock.cpp, which is built as part of
tapyrus_chainstate as before, but also compiled directly into tapyrus-genesis
(with TAPYRUS_GENESIS_SKIP defined to compile out the chain-tip check, along
with its own copy of consensus/tx_verify.cpp) instead of linking
tapyrus_chainstate for CheckBlock(). Static archives extract a whole member
file at once, and checkblock.cpp's chain-tip check touches g_chainstate
(chainstate.cpp) and transitively leveldb, neither of which tapyrus-genesis
needs.

interfaces/node.cpp moves out of tapyrus_util into tapyrus-qt (its only real
consumer, plus test-tapyrus-qt's own copy for its MakeNode() callers) with
SKIP_AUTOUIC set to avoid Qt's AutoUIC misparsing its <ui_interface.h>
include. tapyrus_util no longer needs its conditional tapyrus_wallet link,
so tapyrus-verify/tapyrus-tx/tapyrus-cli/tapyrus-genesis no longer pull
tapyrus_wallet/tapyrus_chainstate/leveldb onto their link line just by
needing tapyrus_util.

Also corrects the library dependency graph comment at the top of
src/CMakeLists.txt, which had drifted out of sync with the actual
target_link_libraries calls (missing tapyrus_util on most entries,
tapyrus_cli incorrectly listed as depending on tapyrus_consensus).

Verified: all four executables now have zero libtapyrus_chainstate.a/
tapyrus_wallet.a/tapyrus_peer.a/tapyrus_netbase.a/leveldb.a on their link
line; full test_tapyrus (554 cases) and test-tapyrus-qt suites pass;
lint-circular-dependencies.sh passes.
