+++ src/qt/CMakeLists.txt
@@ -179,9 +179,10 @@ set_property(SOURCE
   walletmodel.cpp
   walletview.cpp
   sendcoinsdialog.cpp
-  # These three are tapyrus-qt's own glue sources (see target_sources(tapyrus-qt ...)
+  # These four are tapyrus-qt's own glue sources (see target_sources(tapyrus-qt ...)
   # below) -- same ui_interface.h collision, just living in src/ instead of src/qt/.
   ../init.cpp
+  ../interfaces/node.cpp
   ../timeoffsets.cpp
   ../ui_interface.cpp
   PROPERTY SKIP_AUTOUIC ON
@@ -287,9 +288,14 @@ add_windows_resources(tapyrus-qt "${TAPYRUS_QT_RC_FILE}")
 # wires console message-box/notification handlers onto the global
 # uiInterface, which would fight with the Qt handlers tapyrusgui.cpp/
 # tapyrus.cpp register on that same interface.
+#
+# interfaces/node.cpp lives here rather than in tapyrus_util -- see the
+# comment on tapyrus_util's add_library() in the top-level src/CMakeLists.txt
+# for why -- since tapyrus-qt is the only consumer of interfaces/node.h.
 target_sources(tapyrus-qt PRIVATE
   ../file_io.cpp
   ../init.cpp
+  ../interfaces/node.cpp
   ../shutdown.cpp
   ../timeoffsets.cpp
   ../ui_interface.cpp
